support link and license open from right top "help" button
