+++ VMagicMirrorConfig/VMagicMirrorConfig/View/ControlPanelTabs/AboutPanel.xaml
@@ -15,23 +15,13 @@
     <StackPanel HorizontalAlignment="Left"
                 Margin="10">
         <TextBlock Style="{StaticResource NoteText}"
-                        Text="VMagicMirror v1.5.0"/>
-        <TextBlock Style="{StaticResource NoteText}"
                         Text="Copyright (c) 2019-2020 獏星(ばくすたー)"/>
         <TextBlock Style="{StaticResource NoteText}"
                         ToolTip="https://github.com/malaybaku/VMagicMirror">
                 <Hyperlink Click="OnClickHyperLinkToGitHub">Source (GitHub)</Hyperlink>
         </TextBlock>
 
-        <Button Content="{DynamicResource Home_OssLicense}" 
-                HorizontalAlignment="Left"
-                Width="120"
-                Click="OnShowLicenseClick"
-                />
-
-        <Rectangle Width="20" Height="20"
-                   
-                   />
+        <Rectangle Height="20" />
 
 
         <TextBlock Style="{StaticResource NoteText}"
